+++ zed_components/src/zed_camera_one/include/zed_camera_one_component.hpp
@@ -252,6 +252,7 @@ private:
   bool _debugStreaming = false;
   bool _debugAdvanced = false;
   bool _debugNitros = false;
+  bool _debugTf = false;
   // If available, force disable NITROS usage for debugging and testing
   // purposes; otherwise, this is always true.
   bool _nitrosDisabled = false;
@@ -320,6 +321,7 @@ private:
   // <---- Publishers
 
   // ----> Publisher variables
+  bool _usingIPC = false;
   sl::Timestamp _lastTs_grab = 0;  // Used to calculate stable publish frequency
   sl::Timestamp _sdkGrabTS = 0;
   std::atomic<size_t> _colorSubCount;
@@ -447,6 +449,7 @@ private:
   std::unique_ptr<tf2_ros::Buffer> _tfBuffer;
   std::unique_ptr<tf2_ros::TransformListener> _tfListener;
   std::unique_ptr<tf2_ros::TransformBroadcaster> _tfBroadcaster;
+  std::unique_ptr<tf2_ros::StaticTransformBroadcaster> _staticTfBroadcaster;
 
   // Camera IMU transform
   sl::Transform _slCamImuTransf;
@@ -470,6 +473,7 @@ private:
   // <---- Camera infos
 
   // ----> Frame IDs
+  bool _staticImuTfPublished = false;
   std::string _cameraLinkFrameId;
   std::string _cameraCenterFrameId;
   std::string _camImgFrameId;