+++ zed_components/src/zed_camera_one/src/zed_camera_one_component_main.cpp
@@ -25,6 +25,7 @@ using namespace std::placeholders;
 
 namespace stereolabs
 {
+
 ZedCameraOne::ZedCameraOne(const rclcpp::NodeOptions & options)
 : Node("zed_node_one", options),
   _threadStop(false),
@@ -48,30 +49,34 @@ ZedCameraOne::ZedCameraOne(const rclcpp::NodeOptions & options)
   _uptimer(get_clock()),
   _setSvoFrameCheckTimer(get_clock())
 {
+  _usingIPC = options.use_intra_process_comms();
+
   RCLCPP_INFO(get_logger(), "================================");
   RCLCPP_INFO(get_logger(), "    ZED Camera One Component    ");
   RCLCPP_INFO(get_logger(), "================================");
   RCLCPP_INFO(get_logger(), " * namespace: %s", get_namespace());
   RCLCPP_INFO(get_logger(), " * node name: %s", get_name());
+  RCLCPP_INFO(get_logger(), " * IPC: %s", _usingIPC ? "enabled" : "disabled");
   RCLCPP_INFO(get_logger(), "================================");
 
   // Set the name of the main thread for easier identification in
   // system monitors
-  pthread_setname_np(pthread_self(), (get_name() + std::string("_main")).c_str());
+  pthread_setname_np(
+    pthread_self(),
+    (get_name() + std::string("_main")).c_str());
 
-
-  const size_t SDK_MAJOR_REQ = 4;
-  const size_t SDK_MINOR_REQ = 2;
-
-  if (ZED_SDK_MAJOR_VERSION < SDK_MAJOR_REQ ||
-    (ZED_SDK_MAJOR_VERSION == SDK_MAJOR_REQ &&
-    ZED_SDK_MINOR_VERSION < SDK_MINOR_REQ))
+  if (((ZED_SDK_MAJOR_VERSION * 10 + ZED_SDK_MINOR_VERSION) <
+    (SDK_MAJOR_MIN_SUPP * 10 + SDK_MINOR_MIN_SUPP)) ||
+    ((ZED_SDK_MAJOR_VERSION * 10 + ZED_SDK_MINOR_VERSION) >
+    (SDK_MAJOR_MAX_SUPP * 10 + SDK_MINOR_MAX_SUPP)))
   {
     RCLCPP_ERROR_STREAM(
       get_logger(),
       "This version of the ZED ROS2 wrapper is designed to work with ZED SDK "
-      "v" << static_cast<int>(SDK_MAJOR_REQ)
-          << "." << static_cast<int>(SDK_MINOR_REQ) << " or newer.");
+      "v" << static_cast<int>(SDK_MAJOR_MIN_SUPP)
+          << "." << static_cast<int>(SDK_MINOR_MIN_SUPP)
+          << " or newer up to v" << static_cast<int>(SDK_MAJOR_MAX_SUPP)
+          << "." << static_cast<int>(SDK_MINOR_MAX_SUPP) << ".");
     RCLCPP_INFO_STREAM(
       get_logger(), "* Detected SDK v"
         << ZED_SDK_MAJOR_VERSION << "."
@@ -81,6 +86,10 @@ ZedCameraOne::ZedCameraOne(const rclcpp::NodeOptions & options)
     RCLCPP_INFO(get_logger(), "Node stopped. Press Ctrl+C to exit.");
     exit(EXIT_FAILURE);
   }
+
+#if (ZED_SDK_MAJOR_VERSION * 10 + ZED_SDK_MINOR_VERSION) >= 50
+  sl::setEnvironmentVariable("ZED_SDK_DISABLE_PROGRESS_BAR_LOG", "1");
+#endif
 
 // ----> Publishers/Subscribers options
 #ifndef FOUND_FOXY
@@ -595,12 +604,16 @@ void ZedCameraOne::getDebugParams()
     shared_from_this(), "debug.debug_advanced", _debugAdvanced,
     _debugAdvanced, " * Debug Advanced: ");
   sl_tools::getParam(
+    shared_from_this(), "debug.debug_tf", _debugTf, _debugTf,
+    " * Debug TF: ");
+  sl_tools::getParam(
     shared_from_this(), "debug.debug_nitros", _debugNitros,
     _debugNitros, " * Debug Nitros: ");
 
   // Set debug mode
   _debugMode = _debugCommon || _debugVideoDepth || _debugCamCtrl ||
-    _debugSensors || _debugStreaming || _debugAdvanced || _debugNitros;
+    _debugSensors || _debugStreaming || _debugAdvanced ||
+    _debugTf || _debugNitros;
 
   if (_debugMode) {
     rcutils_ret_t res = rcutils_logging_set_logger_level(
@@ -810,6 +823,13 @@ void ZedCameraOne::setupTf2()
   _tfBuffer = std::make_unique<tf2_ros::Buffer>(get_clock());
   _tfListener = std::make_unique<tf2_ros::TransformListener>(*_tfBuffer);
   _tfBroadcaster = std::make_unique<tf2_ros::TransformBroadcaster>(this);
+
+  _staticImuTfPublished = false;
+  if (!_usingIPC) {
+    _staticTfBroadcaster = std::make_unique<tf2_ros::StaticTransformBroadcaster>(this);
+  } else {  // Cannot use TRANSIENT_LOCAL with intra-process comms
+    _staticTfBroadcaster.reset();
+  }
 }
 
 void ZedCameraOne::configureZedInput()
@@ -1681,8 +1701,13 @@ void ZedCameraOne::initPublishers()
   // ----> Camera/imu transform message
   if (_publishSensImuTransf) {
     std::string cam_imu_tr_topic = _topicRoot + "left_cam_imu_transform";
+    auto qos = _qos;
+    if (!_usingIPC) {
+      // Use TRANSIENT_LOCAL durability if not using intra-process comms
+      qos = qos.durability(rclcpp::DurabilityPolicy::TransientLocal);
+    }
     _pubCamImuTransf = create_publisher<geometry_msgs::msg::TransformStamped>(
-      cam_imu_tr_topic, _qos, _pubOpt);
+      cam_imu_tr_topic, qos, _pubOpt);
 
     RCLCPP_INFO_STREAM(
       get_logger(), "  * Advertised on topic: "