+++ CHANGELOG.rst
@@ -26,7 +26,12 @@ LATEST CHANGES
 
 2025-11-24
 ----------
-- Add twist information to the `odom` topic
+- Added twist information to the `odom` topic
+- Added support for the new Virtual Stereo API with SDK v5.1.
+
+  - New launch arguments to setup the virtual camera: `serial_numbers` and `camera_ids`
+  - New `ZedCamera` component parameters to setup the virtual camera: `general.virtual_serial_numbers` and `general.virtual_camera_ids`
+  - **NOTE** ZED MEDIA SERVER IS NO LONGER REQUIRED to create a virtual Stereo camera using two ZED X One cameras.
 
 2025-11-21
 ----------
@@ -35,7 +40,7 @@ LATEST CHANGES
   - From `<camera_name>_left_camera_optical_frame` to `<camera_name>_left_camera_frame_optical`
   - From `<camera_name>_right_camera_optical_frame` to `<camera_name>_right_camera_frame_optical`
   - From `<camera_name>_camera_optical_frame` to `<camera_name>_rgb_camera_frame_optical`
-**NOTE** THIS IS A BREAKING CHANGE. Please update your TF references accordingly.
+  - **NOTE** THIS IS A BREAKING CHANGE. Please update your TF references accordingly.
 
 2025-11-11
 ----------
@@ -45,13 +50,22 @@ LATEST CHANGES
 ----------
 - Fixed `camera_info` publishing when no image topics are subscribed
 
-2025-11-05
+2025-11-09
 ----------
-- Remapped `robot_description` topic to `<camera_name>_description` to allow multi-camera URDF integration
+- Added debug option for TF broadcasting
+
+  - Improved TF debug logs to show frame transformations when enabled
+
+- Static baseline information from URDF is now overwritten by the real baseline value retrieved from the camera calibration file.
+- Removed mandatory `custom_baseline` launch argument for virtual stereo cameras made with two ZED X One cameras.
+  The value is retrieved from the calibration file.
+- IMU TF is now broadcast as static if IPC is disabled.
+- IMU Transform topic is now published with TRANSIENT LOCAL durability if IPC is disabled.
 
 2025-11-05
 ----------
-- Changed minimum depth value to 0.01 meters when using ZED SDK v5.1 or higher 
+- Remapped `robot_description` topic to `<camera_name>_description` to allow multi-camera URDF integration
+- Changed minimum depth value to 0.01 meters when using ZED SDK v5.1 or higher
 
 2025-10-31
 ----------